requirements.txt
fastapi
uvicorn
python-dotenv
requests
pydantic
# LangChain & AI
langchain
langchain-openai
langchain-community
langgraph
# Azure
azure-search-documents
azure-core
# Twilio
twilio